Improve logon with 2FA
Currenly when trying to autenticate against a CyberArk with 2FA enabled the
request request fails with the following error:

    Failed to Logon to the PVWA. Failed to authenticate to the PAS REST API.
    Received non-200 status code '500'

In function `httpjson.SendRequestRay` the response body isn't returned, so
the caller function never finds the correct error code.

After the previous fix, the second call, the one with the otp code still fails
with 403. Debugging it seems that we need to store the cookies between the
two logon calls. So a go context is added to the functions.