Improve logon with 2FA (#114)
* V0.1.3 beta (#113)

* Fixes #108 add authn ldap support (#109)
* Fixes #106 Adding 6 CEM sub-commands (#105) (#111)

Co-authored-by: Quincy Cheng <quincy.cheng@gmail.com>
Co-authored-by: Joe Garcia <infamousjoeg@users.noreply.github.com>

* Improve logon with 2FA

Currenly when trying to autenticate against a CyberArk with 2FA enabled the
request request fails with the following error:

    Failed to Logon to the PVWA. Failed to authenticate to the PAS REST API.
    Received non-200 status code '500'

In function `httpjson.SendRequestRay` the response body isn't returned, so
the caller function never finds the correct error code.

After the previous fix, the second call, the one with the otp code still fails
with 403. Debugging it seems that we need to store the cookies between the
two logon calls. So a go context is added to the functions.

* Fix lint errors and add missing file from previous commit

This commit fixes the go  lint error related to key type in context and it
adds a missing file from previous commit not uploaded by error.

* Fix type error in comment

* Missing comment on Cookie function

Added missing comment

* Fix panic when body response is nil

Try to catch nil response body on httpjson.SendRequestRaw function

Co-authored-by: Andrew Copeland <50109276+AndrewCopeland@users.noreply.github.com>
Co-authored-by: Quincy Cheng <quincy.cheng@gmail.com>
Co-authored-by: Joe Garcia <infamousjoeg@users.noreply.github.com>